fixed bug with debit cards

+++ src/bankaccountapp/Checking.java
@@ -3,7 +3,7 @@ package bankaccountapp;
 public class Checking extends Account {
 
 	// List properties specific to a checking account
-	private int debitCardNumber;
+	private long debitCardNumber;
 	private int debitCardPin;
 	
 	// Constructor to initialize checking account properties
@@ -15,7 +15,7 @@ public class Checking extends Account {
 	
 	// List any methods specific to checking account
 	private void setDebitCard() {
-		this.debitCardNumber = (int) (Math.random() * Math.pow(10,  12));
+		this.debitCardNumber = (long) (Math.random() * Math.pow(10,  12));
 		this.debitCardPin = (int) (Math.random() * Math.pow(10,  4));
 	}
